backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.29.7
twilio>=8.10.0
django-filter==23.4
# django-ratelimit==4.1.0  # Using custom rate limiting
bleach==6.1.0
celery==5.3.4
redis==5.0.1
